@@ -1,6 +1,6 @@
 /// Database authentication information.
 public static class DatabaseAuth
 {
-    public static string DB_USERNAME = "neo4j";
-    public static string DB_PASSWORD = "bigdata2019";
+    public static string DB_USERNAME = "neo4j"; // Local database username
+    public static string DB_PASSWORD = "bb//database"; // Local database password
 }
